Изменил на новое ТЗ

+++ routes/index.js
@@ -8,55 +8,64 @@ let util = require('util');
 let fileReaderNew = util.promisify(fs.readFile);
 
 let busboy = require('async-busboy');
-let lodash = require('lodash');
-let AuthService = require('../service/AuthService');
-let TabService = require('../service/TabService');
-let PagesService = require('../service/PagesService');
-let PostsService = require('../service/PostsService');
-let UtilsService = require('../service/UtilsService');
-let MsqlService = require('../service/MsqlService');
-let GosProgrammServise = require('../service/GosProgrammServise');
-let router = express.Router();
-
-const parse = require('csv-parse');
 
 const XLSX = require('xlsx');
-let parceCVS = util.promisify(parse);
-
 
 
 const jsonwebtoken = require('jsonwebtoken');
 
 
 
+let AuthService = require('../service/AuthService');
+let checkSession = require('../utils/checkSession');
 
 
-router.post("/getgosprogramm", async(req, res, next) => {
-
-
-    let result = await GosProgrammServise.addPost(req.body);
-
-
-
-
-
-    res.json({code: 0});
+let PostsService = require('../service/PostsService');
+let UtilsService = require('../service/UtilsService');
+let MsqlService = require('../service/MsqlService');
+let MenuService = require('../service/MenuService');
 
 
 
 
 
 
-});
+let router = express.Router();
+
+
+
+
+
+
+
+
+
+
+
 
 
 router.get("/getallroles", async (req, res, next) => {
 
+    if (await checkSession(req, "read")){
 
-    let result = await AuthService.getAllRoles();
 
 
-    res.json({code: 0, resultFromDB: result});
+        let result = await AuthService.getAllRoles();
+
+
+        res.json({code: 0, resultFromDB: result});
+
+
+
+    } else {
+
+
+        res.json({code: 1});
+
+
+
+    }
+
 
 
 });
@@ -64,15 +73,28 @@ router.get("/getallroles", async (req, res, next) => {
 
 router.post("/adduser", async(req, res, next) => {
 
-
-
-    let result = await AuthService.addUsers(req.body);
-
+    if (await checkSession(req, "AuthService")){
 
 
 
+        let result = await AuthService.addUsers(req.body);
 
-    res.json({code: 0, resultFromDB: result.ops[0]._id});
+
+
+
+
+        res.json({code: 0, resultFromDB: result.ops[0]._id});
+
+
+
+
+    } else {
+
+        res.json({"code": 1});
+
+
+    }
+
 
 
 
@@ -85,10 +107,22 @@ router.post("/adduser", async(req, res, next) => {
 router.get("/getallusers", async (req, res, next) => {
 
 
-    let result = await AuthService.getAllUsers();
+    if (await checkSession(req, "read")) {
 
 
-    res.json({code: 0, resultFromDB: result.slice(2)});
+        let result = await AuthService.getAllUsers();
+
+
+        res.json({code: 0, resultFromDB: result.slice(2)});
+
+
+    } else {
+
+
+
+        res.json({"code": 1});
+
+    }
 
 
 });
@@ -98,12 +132,22 @@ router.post("/updateusers", async(req, res, next) => {
 
 
 
-    let result = await AuthService.updUsers(req.body);
+    if (await checkSession(req, "AuthService")) {
 
 
-    res.json({code: 0, resultFromDB: result});
+        let result = await AuthService.updUsers(req.body);
 
 
+        res.json({code: 0, resultFromDB: result});
+
+    } else {
+
+
+
+        res.json({"code": 1});
+
+
+    }
 
 
 
@@ -115,12 +159,20 @@ router.post("/deleteoneusers", async(req, res, next) => {
 
 
 
-    let result = await UtilsService.deleteByIdCommon(req.body._id, "users");
+    if (await checkSession(req, "AuthService")) {
 
 
-    res.json({code: 0, resultFromDB: result});
+        let result = await UtilsService.deleteByIdCommon(req.body._id, "users");
 
 
+        res.json({code: 0, resultFromDB: result});
+
+    } else {
+
+
+        res.json({"code": 1});
+
+    }
 
 
 
@@ -183,27 +235,43 @@ router.post('/auth', async(req, res, next) => {
 router.get("/getalltable", async(req, res, next) => {
 
 
-    let result = await MsqlService.getTables();
-    let resultAllTable = [];
 
 
+    if (await checkSession(req, "read")) {
 
-    if(result.hasOwnProperty("recordset")) {
 
-        for (let itemTable of result.recordset) {
+        let result = await MsqlService.getTables();
+        let resultAllTable = [];
 
-           resultAllTable.push(itemTable.TABLE_NAME);
+
+        if (result.hasOwnProperty("recordset")) {
+
+            for (let itemTable of result.recordset) {
+
+                resultAllTable.push(itemTable.TABLE_NAME);
+
+            }
 
         }
 
+
+
+
+
+
+        res.json({code: 0, resultFromDB: resultAllTable});
+
+        resultAllTable = [];
+        result = [];
+
+
+    } else {
+
+
+        res.json({"code": 1});
+
+
     }
-
-
-
-    res.json({code: 0, resultFromDB: resultAllTable});
-
-    resultAllTable = [];
-    result = [];
 
 
 });
@@ -214,29 +282,33 @@ router.get("/getalltable", async(req, res, next) => {
 router.post("/getdatafrommssql", async(req, res, next) => {
 
 
+    if (await checkSession(req, "read")) {
 
 
-    let result = await MsqlService.getDataFromOneTable(req.body.tableName);
+        let result = await MsqlService.getDataFromOneTable(req.body.tableName);
 
 
+        if (result.hasOwnProperty("recordset")) {
+
+            res.json({code: 0, resultFromDB: result.recordset});
 
 
-    if(result.hasOwnProperty("recordset")) {
+        } else {
 
-        res.json({code: 0, resultFromDB: result.recordset});
+            res.json({code: 1});
 
+        }
+
+        result = [];
 
 
     } else {
 
-        res.json({code: 1});
+
+        res.json({"code": 1});
+
 
     }
-
-    result =  [];
-
-
-
 
 
 });
@@ -246,23 +318,29 @@ router.post("/getdatafrommssql", async(req, res, next) => {
 
 router.post("/deleteonetable", async(req, res, next) => {
 
+    if (await checkSession(req, "MsqlService")) {
 
 
-    let result = await MsqlService.deleteOneTable(req.body.tableName);
+        let result = await MsqlService.deleteOneTable(req.body.tableName);
 
 
-   if(result.recordset === undefined){
+        if (result.recordset === undefined) {
 
-       res.json({code: 0});
+            res.json({code: 0});
 
-   } else {
+        } else {
 
-       res.json({code: 1});
+            res.json({code: 1});
 
-   }
+        }
 
 
+    } else {
 
+
+        res.json({"code": 1});
+
+    }
 
 
 
@@ -273,77 +351,34 @@ router.post("/deleteonetable", async(req, res, next) => {
 router.post("/insertnewdata", async(req, res, next) => {
 
 
-    await MsqlService.insertNewData(req.body.tableName, req.body.changes[0]);
 
 
-
-    res.json({code: 0});
-
+    if (await checkSession(req, "MsqlService")) {
 
 
+        await MsqlService.insertNewData(req.body.tableName, req.body.changes[0]);
 
 
-
-});
-
+        res.json({code: 0});
 
 
-
-router.get("/getallpages", async(req, res, next) => {
-
-
-    let arrResult = [];
-
-    let result = await PagesService.getAllPages();
-    let resultStructure = await PagesService.getAllStructure();
-    let resultIspolStructure = await PagesService.getAllIspolStructure();
+    } else {
 
 
-    arrResult.push(result);
-    arrResult.push(resultStructure);
-    arrResult.push(resultIspolStructure);
+        res.json({"code": 1});
 
-    res.json({code: 0, resultFromDB: arrResult});
-
-
+    }
 
 
 
 });
 
 
-router.get("/getallcharts", async(req, res, next) => {
 
 
 
 
 
-    let result = await PagesService.getAllCharts();
-
-
-    res.json({code: 0, resultFromDB: result});
-
-
-
-
-
-});
-
-
-router.get("/getalloptions", async(req, res, next) => {
-
-
-
-    let result = await PagesService.getAllOptions();
-
-
-    res.json({code: 0, resultFromDB: result});
-
-
-
-
-
-});
 
 
 router.post("/checkadminpage", async(req, res, next) => {
@@ -373,297 +408,34 @@ router.post("/checkadminpage", async(req, res, next) => {
 
 
 
-
-
-function getRandomInt(min, max) {
-    return Math.floor(Math.random() * (max - min)) + min;
-}
-
-
-
-
-
-router.get("/dynamicpage", async(req, res, next) => {
-
-
-
-
-
-
-
-
-    let resultFromDB = await PostsService.getOnePost(req.query.id);
-
-
-
-
-    let groupTab = lodash.groupBy(resultFromDB, "tabName");
-
-    groupTab = Object.keys(groupTab);
-
-
-
-
-
-    let pageType = await PagesService.getPageType(req.query.id);
-
-
-
-     let dataFromMssql = [];
-
-     let randomPrefix = getRandomInt(1, 1000000);
-
-
-
-
-
-     for (let [index, obj] of resultFromDB.entries()) {
-
-
-
-         dataFromMssql.push(await MsqlService.getDataFromOneTable(obj.tableName));
-
-
-         await PagesService.insertFromMSSQL(dataFromMssql[index].recordset, obj.tableName + randomPrefix);
-
-
-
-
-
-
-     }
-
-
-
-
-
-
-
-    let resultAllTable = [];
-
-
-
-
-
-
-
-
-         for (let [index, itemResult] of resultFromDB.entries()) {
-             let result = [];
-
-             let tabName = await TabService.getById(itemResult.tabName);
-
-
-
-
-
-
-
-             let chart = await PagesService.getChartById(itemResult.chartId);
-
-
-
-
-             if (chart.type === "pie" || chart.type === "donut") {
-
-                 result.push(await PagesService.getDataPie(itemResult.tableName + randomPrefix, itemResult.chipsArr, itemResult.typeDiagramm));
-
-
-
-                 resultAllTable.push({"typeDiagramm": itemResult.typeDiagramm, "idElemTab": slug(tabName.tabName, {lower: true}), "dataRow": dataFromMssql[index], "titleCharts": itemResult.titleCharts, "data": result, "categ": itemResult.chipsArr, "description": itemResult.description, "fileName": itemResult.fileName, "fileUrl": itemResult.fileUrl, "chartType": chart.type, "idElem": slug(itemResult.titleCharts, {lower: true}), "typePage": pageType.type, "axisRotate": chart.axisRotate, "stackBar": chart.stackBar});
-
-
-                 await PagesService.deleteTempTable(itemResult.tableName + randomPrefix);
-
-
-
-             } else {
-
-
-
-
-
-                 result.push(await PagesService.getDataForDynamicPage(itemResult.tableName + randomPrefix, itemResult.chipsArr, itemResult.typeDiagramm));
-
-
-
-
-
-
-
-
-
-                 if (tabName === 0){
-
-
-                     console.log("\x1b[42m", 0);
-                 } else {
-
-                     resultAllTable.push({"typeDiagramm": itemResult.typeDiagramm, "idElemTab": slug(tabName.tabName, {lower: true}), "dataRow": dataFromMssql[index], "titleCharts": itemResult.titleCharts, "data": result, "categ": itemResult.chipsArr, "description": itemResult.description, "fileName": itemResult.fileName, "fileUrl": itemResult.fileUrl, "chartType": chart.type, "idElem": slug(itemResult.titleCharts, {lower: true}), "typePage": pageType.type, "axisRotate": chart.axisRotate, "stackBar": chart.stackBar});
-
-                 }
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-                 await PagesService.deleteTempTable(itemResult.tableName + randomPrefix);
-
-
-
-
-             }
-
-
-
-
-
-
-
-
-
-
-         }
-
-
-
-
-
-
-             let tabName = [];
-
-
-             for (let groupTabOne of groupTab) {
-
-
-                 let tempObj = await TabService.getById(groupTabOne);
-
-                 if (tempObj !== 0) {
-
-                     tempObj["idElemTab"] = slug(tempObj.tabName, {lower: true});
-
-                     tabName.push(tempObj);
-
-                 }
-
-
-
-
-
-             }
-
-
-
-
-
-
-
-
-
-
-
-        res.json({code: 0, resultFromDB: resultAllTable, tabNames: tabName});
-
-
-
-
-
-
-
-
-
-
-
-});
-
-
-
-
 router.post("/addpost", async(req, res, next) => {
 
 
-    const {files, fields} = await busboy(req);
+    if (await checkSession(req, "PostsService")) {
 
 
-
-    let pathForWrite = path.join(__dirname, process.env.PATHUPLOAD);
-    fields["fileNameArr"] = [];
-    fields["fileUrlArr"] = [];
-
-    for (let fileItem of files) {
+        let result = await PostsService.addPost(req.body);
 
 
+        if (result.result.ok === 1) {
 
-        fields.fileNameArr.push(fileItem.filename);
-        fields.fileUrlArr.push(`uploads/${path.basename(fileItem.path)}`);
-
-        fileItem.pipe(fs.createWriteStream(pathForWrite + path.basename(fileItem.path)));
-
-    }
+            res.json({"code": 0});
 
 
-
-    let result = await PostsService.addPost(fields);
-
+        } else {
 
 
-    if (result.hasOwnProperty("result")){
+            res.json({"code": 1});
 
-        res.json({code: 0});
+
+        }
+
 
     } else {
-        res.json({code: 1});
-    }
 
 
 
-
-
-
-
-
-
-
-
-});
-
-
-
-router.post("/addcvs", async(req, res, next) => {
-
-
-    const {files, fields} = await busboy(req);
-
-
-    let records = [];
-    let tableName = [];
-
-
-    for (let fileItem of files) {
-
-
-
-
-        let tempStr = await fileReaderNew(fileItem.path);
-
-
-
-
-         records.push(await parceCVS(tempStr.toString(), {columns: true, delimiter: ";"}));
-         tableName.push(path.parse(fileItem.filename).name);
-
-
-
+        res.json({"code": 1});
 
 
 
@@ -671,52 +443,41 @@ router.post("/addcvs", async(req, res, next) => {
 
 
 
-
-
-    await MsqlService.insertOneTable(tableName, records);
-
-
-
-
-
-
-
-   res.json({code: 0});
-
-
-
-
-
-
 });
 
 
-router.post("/getgosprogramm", async(req, res, next) => {
-
-
-   let result = await GosProgrammServise.addPost(req.body);
 
 
 
-
-
-   res.json({code: 0});
-
-
-
-
-
-
-});
 
 
 router.post("/deleteonepost", async (req, res, next) => {
 
 
-    let result = await PostsService.deleteOnePost(req.body.id);
+
+    if (await checkSession(req, "PostsService")) {
 
 
-    res.json({code: 0});
+        let result = await PostsService.deleteOnePost(req.body.id);
+
+        if (result.result.ok === 1) {
+
+            res.json({"code": 0});
+
+        } else {
+
+            res.json({"code": 1});
+
+        }
+
+
+    } else {
+
+
+        res.json({"code": 1});
+
+
+    }
 
 
 });
@@ -724,92 +485,197 @@ router.post("/deleteonepost", async (req, res, next) => {
 router.get("/getallpost", async (req, res, next) => {
 
 
-    let result = await PostsService.getAllPost();
+    if (await checkSession(req, "read")) {
 
 
-   res.json({code: 0, resultFromDB: result});
+        let result = await PostsService.getAllPost();
 
 
-});
-
-router.post("/sendnewtabname", async(req, res, next) => {
+        res.json({code: 0, resultFromDB: result});
 
 
+    } else {
 
-    let result = await TabService.createNewTab(req.body.tabName);
+        res.json({"code": 1});
 
-
-
-    res.json({code: 0, resultFromDB: result.ops[0]});
-
-
-
-
+    }
 
 });
 
 
-router.get("/getalltabs", async(req, res, next) => {
-
-
-    let result = await TabService.getAll();
-
-
-
-    res.json({code: 0, resultFromDB: result});
 
 
 
 
 
+
+router.get('/getonepost', async (req, res, next) =>{
+
+
+    let result = await PostsService.getOnePost(req.query.id);
+
+
+
+
+    res.json({"code": 0, "resultFromDb": result});
 
 
 });
+
+
+
 
 
 router.post("/addexcel", async(req, res, next) => {
 
 
-    const {files, fields} = await busboy(req);
+
+    if (await checkSession(req, "MsqlService")) {
 
 
-    let records = [];
+        const {files, fields} = await busboy(req);
 
 
-    for (let fileItem of files) {
+        let records = [];
 
 
-        const workbook = XLSX.readFile(fileItem.path);
-        let tableName = path.parse(fileItem.filename).name;
+        for (let fileItem of files) {
 
 
-
-        let allList = workbook.SheetNames;
-
-
-
-        let onlyFirst = workbook.Sheets[allList[0]];
+            const workbook = XLSX.readFile(fileItem.path);
+            let tableName = path.parse(fileItem.filename).name;
 
 
-        let allArrResult = XLSX.utils.sheet_to_json(onlyFirst);
+            let allList = workbook.SheetNames;
 
 
-        records = allArrResult;
+            let onlyFirst = workbook.Sheets[allList[0]];
 
 
-        await MsqlService.insertOneTable(tableName, records);
+            let allArrResult = XLSX.utils.sheet_to_json(onlyFirst);
 
-        records = [];
+
+            records = allArrResult;
+
+
+            await MsqlService.insertOneTable(tableName, records);
+
+            records = [];
+        }
+
+
+        res.json({code: 0});
+
+
+    } else {
+
+
+        res.json({"code": 1});
+
+
     }
 
 
 
 
 
+});
 
-    res.json({code: 0});
 
 
+function setSlugForTree(menusArr) {
+
+
+    menusArr.childs = menusArr.childs
+        .map(function (child) {
+
+
+            child["url"] = slug(child.title, {lower: true});
+
+
+            return child;
+
+
+        }).map(function (child) {
+            return setSlugForTree(child);
+        });
+
+
+    return menusArr;
+
+}
+
+
+
+router.post('/savemenu', async (req, res, next) =>{
+
+
+
+
+    if (await checkSession(req, "MenuService")) {
+
+
+        let clearArr = [];
+
+
+        for (let itemOneMenuObj of req.body.menus) {
+            let tempArr = itemOneMenuObj;
+
+
+            clearArr.push(setSlugForTree(tempArr));
+
+            tempArr = [];
+
+        }
+
+
+        for (let itemOneParentObj of clearArr) {
+            itemOneParentObj["url"] = slug(itemOneParentObj.title, {lower: true});
+
+        }
+
+
+        let result = await MenuService.addMenus(clearArr);
+
+
+        if (result.result.ok === 1) {
+
+            res.json({"code": 0});
+
+
+        } else {
+
+
+            res.json({"code": 1});
+
+
+        }
+
+    } else {
+
+        res.json({"code": 1});
+
+
+    }
+
+
+
+});
+
+
+
+router.get('/allmenus', async (req, res, next) =>{
+
+
+
+
+
+
+
+        let result = await MenuService.getAllMenus();
+
+
+        res.json({"code": 0, "resultFromDb": result});
 
 
 
@@ -818,6 +684,8 @@ router.post("/addexcel", async(req, res, next) => {
 
 
 });
+
+
